refactor(e2e): address Paola's review comments on check_failure test
- Use environments.Host + awshost.Provisioner instead of custom env/provisioner
- Embed Python check and YAML config fixtures via go:embed
- Fix race conditions: restart agent before flushing fakeintake in Phase 2 & 3
- Remove redundant agent-ready wait (framework guarantees readiness at suite start)
