+++ test/new-e2e/tests/agent-health/fixtures/broken_check.py
@@ -1,0 +1,6 @@
+from datadog_checks.base import AgentCheck
+
+
+class BrokenCheck(AgentCheck):
+    def check(self, instance):
+        raise Exception("synthetic failure for e2e health platform test")
