+++ test/new-e2e/tests/agent-health/fixtures/broken_check.yaml
@@ -1,0 +1,3 @@
+init_config:
+instances:
+  - {}
