+++ test/new-e2e/tests/agent-health/fixtures/fixed_check.py
@@ -1,0 +1,6 @@
+from datadog_checks.base import AgentCheck
+
+
+class BrokenCheck(AgentCheck):
+    def check(self, instance):
+        self.gauge("e2e.healthplatform.check_ok", 1)
